+++ packages/vue/src/components/atoms/SfSticky/SfSticky.html
@@ -1,3 +1,3 @@
 <div class="sf-sticky">
-    <slot/>
+  <slot/>
 </div>
