+++ packages/vue/src/components/atoms/SfSticky/SfSticky.spec.ts
@@ -1,5 +1,5 @@
 import { shallowMount } from "@vue/test-utils";
-import SfSticky from "@/components/molecules/SfSticky.vue";
+import SfSticky from "@/components/atoms/SfSticky.vue";
 
 describe("SfSticky.vue", () => {
   it("renders a component", () => {
